requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# AI/ML
openai==1.3.7
diffusers==0.24.0
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1

# File Upload & Processing
python-multipart==0.0.6
aiofiles==23.2.1

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-magic==0.4.27

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: For better performance
orjson==3.9.10 